feat: Add Support for iOS 17 (#11)
Considering the changes introduced in EventKit for iOS 17, add the "requestFullAccessToEventWithCompletion:" method to trigger the alert to allow full access to the Calendar. This allows the plugin to keep a similar behaviour to the 16 version (and previous). Considering that the methods use the same access grant closure, this was moved into a local variable that can be called by all callers.
The iOS 5 condition was removed as our current supported minimum version is higher than that one.
Add the "NSCalendarsFullAccessUsageDescription" info.plist key added by Apple for iOS 17. It uses the same text set for "NSCalendarsUsageDescription", keeping the same experience as before.
Remove the SYSTEM_VERSION_GREATER_THAN_OR_EQUAL_TO macro as it's not used in the code.

References: https://outsystemsrd.atlassian.net/browse/RMET-2726